Added set_beam_mat() that makes the attribute beam_mat - just the beam part of the A-matrix.
Added set_phase_mat() that makes the attribute phase_mat - just the phase
part of the A-matrix.

Removed the calculation of the beam_mat attribute from set_a_mat().
So the user will need to call set_beam_mat() to get the beam matrix;
it will no longer be done by set_a_mat().

The motivation for these changes is to have access to the two components
of the A-matrix separately, and not have to store all three matrices in memory.