horizon_map.py:
Changed the name of the key for the beam weight map to 'beam_weight_sum' to be
consistent with image power spectrum pipeline.

The image PS pipeline uses the squared beam weights for normalization.  Added
the computation of this, and its key is 'beam_sq_weight_sum' in the map
dictionary, again to be consistent with what image PS expects.

Added two paramaters to the class:  return_b1map and return_b2map, both
with default value False.  The two beam weight maps get loaded into the
map dictionary only if the user sets the parameters to True.  When they
are False, only the normalized map gets loaded into the map dictionary.

dom_pspec.py:

This seems to expect the maps to be in 'unnormalized' units.  However,
other units are possible and we should allow for that.  So I added the
parameter units to the class with default value 'unnormalized'.

Changed the import of hera_pspec.conversions to from direct_optimal_mapping
import conversions.  The latter was set up by Vincent MacKay to allow us to
avoid using hera_pspec.conversions.  Importing hera_pspec.conversions brings
in a lot of baggage, like all the hera calibration modules.